implemeted a dedicated component for bed capacity in facility home
